Move TokensModel to domain-layer
